:recycle: Reorganize exposed function and filenames
