+++ packages/lisk-transactions/src/index.ts
@@ -32,30 +32,76 @@ import { createSignatureObject } from './create_signature_object';
 import { TransactionError, TransactionMultiError } from './errors';
 import { Status, TransactionResponse } from './response';
 import { TransactionJSON } from './transaction_types';
-import * as utils from './utils';
+import {
+	checkPublicKeysForDuplicates,
+	convertBeddowsToLSK,
+	convertLSKToBeddows,
+	getTransactionBytes,
+	getTransactionHash,
+	getTransactionId,
+	isValidInteger,
+	multiSignTransaction,
+	prepareTransaction,
+	prependMinusToPublicKeys,
+	prependPlusToPublicKeys,
+	signRawTransaction,
+	signTransaction,
+	validateAddress,
+	validateFee,
+	validateKeysgroup,
+	validatePublicKey,
+	validatePublicKeys,
+	validateTransaction,
+	verifyTransaction,
+} from './utils';
+
+const exposedUtils = {
+	convertBeddowsToLSK,
+	convertLSKToBeddows,
+	isValidInteger,
+	multiSignTransaction,
+	prependMinusToPublicKeys,
+	prependPlusToPublicKeys,
+	validateAddress,
+	validateKeysgroup,
+	validatePublicKey,
+	validatePublicKeys,
+
+	// TODO: Deprecated
+	signTransaction,
+	getTransactionBytes,
+	getTransactionId,
+	verifyTransaction,
+	checkPublicKeysForDuplicates,
+	getTransactionHash,
+	prepareTransaction,
+	signRawTransaction,
+	validateFee,
+	validateTransaction,
+};
 
 export {
 	BaseTransaction,
-	castVotes,
-	createDapp,
-	MultisignatureTransaction,
 	TransferTransaction,
-	DappTransaction,
+	transfer,
 	SecondSignatureTransaction,
+	registerSecondPassphrase,
 	DelegateTransaction,
+	registerDelegate,
 	VoteTransaction,
+	castVotes,
+	MultisignatureTransaction,
+	registerMultisignature,
+	DappTransaction,
+	createDapp,
 	InTransferTransaction,
 	OutTransferTransaction,
-	transfer,
-	registerDelegate,
-	registerMultisignature,
-	registerSecondPassphrase,
 	createSignatureObject,
-	utils,
-	constants,
 	Status,
 	TransactionResponse,
 	TransactionJSON,
 	TransactionError,
 	TransactionMultiError,
+	constants,
+	exposedUtils as utils,
 };
