+++ packages/lisk-transactions/test/utils/verify.ts
@@ -21,7 +21,7 @@ import {
 	verifySecondSignature,
 	verifyMultiSignatures,
 } from '../../src/utils';
-import * as validator from '../../src/utils/sign_and_verify';
+import * as validator from '../../src/utils/sign_and_validate';
 import { TransactionError, TransactionPendingError } from '../../src/errors';
 import { MultisignatureStatus } from '../../src/base_transaction';
